Criando a pasta produtos

+++ application/controllers/Produtos.php
@@ -1252,6 +1252,7 @@ class Produtos extends CI_Controller {
 		
         $data['file'] = $this->input->post(array(
             'idTab_Produto',
+			'idSis_Empresa',
             'Arquivo',
 		), TRUE);
 
@@ -1342,6 +1343,7 @@ class Produtos extends CI_Controller {
 				endswitch;			
 
                 $data['camposfile'] = array_keys($data['file']);
+				$data['file']['idSis_Empresa'] = $_SESSION['Empresa']['idSis_Empresa'];
 				$data['idSis_Arquivo'] = $this->Produtos_model->set_arquivo($data['file']);
 
                 if ($data['idSis_Arquivo'] === FALSE) {
@@ -1366,10 +1368,14 @@ class Produtos extends CI_Controller {
 						exit();
 					} else {
 
-						if(null!==('arquivos/imagens/empresas/' . $_SESSION['Empresa']['idSis_Empresa'] . '/produtos/original/' . $_SESSION['Produtos']['Arquivo'] . '')){
+						if((null!==('arquivos/imagens/empresas/' . $_SESSION['Empresa']['idSis_Empresa'] . '/produtos/original/' . $_SESSION['Produtos']['Arquivo'] . ''))
+							&& (('arquivos/imagens/empresas/' . $_SESSION['Empresa']['idSis_Empresa'] . '/produtos/original/' . $_SESSION['Produtos']['Arquivo'] . '')
+							!==('arquivos/imagens/empresas/' . $_SESSION['Empresa']['idSis_Empresa'] . '/produtos/original/fotoproduto.jpg'))){
 							unlink('arquivos/imagens/empresas/' . $_SESSION['Empresa']['idSis_Empresa'] . '/produtos/original/' . $_SESSION['Produtos']['Arquivo'] . '');						
 						}
-						if(null!==('arquivos/imagens/empresas/' . $_SESSION['Empresa']['idSis_Empresa'] . '/produtos/miniatura/' . $_SESSION['Produtos']['Arquivo'] . '')){
+						if((null!==('arquivos/imagens/empresas/' . $_SESSION['Empresa']['idSis_Empresa'] . '/produtos/miniatura/' . $_SESSION['Produtos']['Arquivo'] . ''))
+							&& (('arquivos/imagens/empresas/' . $_SESSION['Empresa']['idSis_Empresa'] . '/produtos/miniatura/' . $_SESSION['Produtos']['Arquivo'] . '')
+							!==('arquivos/imagens/empresas/' . $_SESSION['Empresa']['idSis_Empresa'] . '/produtos/miniatura/fotoproduto.jpg'))){
 							unlink('arquivos/imagens/empresas/' . $_SESSION['Empresa']['idSis_Empresa'] . '/produtos/miniatura/' . $_SESSION['Produtos']['Arquivo'] . '');						
 						}
